Allow piping to cvlc or ffmpeg (#214)
It was already possible to pipe the frames on stdout as RGB. This PR intends to generalize the output color space so that we can choose which one to output. Since internally our frames are in BGR format, we use it now as default. Added --output_color_space to change it.
See README for me example on how to use it.